Swap LBAs for both drives going into vdrive
since it seems that the vdrive implementation swaps them itself.
The symptom was that drive 0 always got the data from track 18, no
matter which track you wanted to read. This was because drive 1: was not
used and kept its head in track 18. The data was from the correct disk
though.

Also:
Supply core clock to drive for output bus buffering.
Tweak the drive menu  a bit.